requirements.txt
Django==2.2.14
django-allauth==0.40.0
django-environ
django-debug-toolbar